+++ src/code/waves/waves.hpp
@@ -6,8 +6,8 @@ struct Wave
 {
     int id;
     std::vector<Enemy> enemies_in_wave;
-    int start;
-    int freq_btw_ennemies_in_s;
+    double start;
+    double freq_btw_ennemies_in_s;
 
     int nbr_knights;
     int nbr_wizards;
